修复信息刮削bug，文件扫描bug，需要测试

+++ middlewars/info_completion.py
@@ -5,12 +5,13 @@ from requests import get
 from os import path,getcwd
 from traceback import format_exc
 from datetime import datetime
+from const import *
 from threading import Thread
 
 from loguru import logger
 
+
 from utils import Sqlite_con,MysqlCon,YamlConfig
-from const import *
 
 
 class InfoCompletion:
@@ -18,18 +19,19 @@ class InfoCompletion:
         self.base_url = "http://git.znana.top:4000"
     
     # 异步请求
-    async def get_info(self,info:dict,url:str,sem:asyncio.Semaphore):
-        async with sem:
-            async with httpx.AsyncClient()as cli:
-                try:
-                    res = await cli.get(url=url)
-                    return info,res
-                except:
-                    logger.error(format_exc())
-                    return info,None
+    async def get_info(self,info:dict,url:str):
+        async with httpx.AsyncClient()as cli:
+            try:
+                res = await cli.get(url=url)
+                return info,res
+            except:
+                logger.error(format_exc())
+                return info,None
     # 歌曲照片保存
     def save_img(self,pic_name:str,url:str) -> str:
         img_path = path.join(getcwd(),"data","album_img",pic_name+".jpg")
+        if path.exists(img_path):
+            return img_path
         res = get(url=url)
         with open(img_path,"wb")as f:
             f.write(res.content)
@@ -37,6 +39,8 @@ class InfoCompletion:
     # 歌曲歌词保存
     def save_lyric(self,lrc_name:str,netease_id:str) -> str:
         lrc_path = path.join(getcwd(),"data","lrcs",lrc_name+".lrc")
+        if path.exists(lrc_path):
+            return lrc_path
         lyric_url = self.base_url+f"/lyric/new?id={netease_id}"
         res = get(lyric_url).json()
         lrc = res.get("lrc").get("lyric")
@@ -56,6 +60,7 @@ class InfoCompletion:
                 mid = info.get("id")
                 title = info.get("title")
                 album = info.get("album")
+                pic_url = None
                 for n in res.get("result").get("songs"):
                     name = n.get("name")
                     ar_name = "|".join([x.get("name") for x in n.get("ar")])
@@ -69,26 +74,30 @@ class InfoCompletion:
                     publish_time = n.get("publishTime")
                     netease_id = n.get("id")
                     break
-                img_path = self.save_img(title+"-"+album,pic_url)
-                lrc_path = self.save_lyric(ar_name+"-"+album+"-"+title,netease_id=netease_id)
-                curdate = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                media_sql = f'update media set netease_id="{netease_id}",artist_name="{ar_name}",image_path="{img_path}",lyric_path="{lrc_path}",update_at="{curdate}" where id="{mid}";'
-                artist_sql = f'insert or ignore into artist(name) values("{ar_name}");'
-                sql_con.sql2commit(media_sql)
-                sql_con.sql2commit(artist_sql)
+                if pic_url:
+                    img_path = self.save_img(title+"-"+album,pic_url)
+                    lrc_path = self.save_lyric(ar_name+"-"+album+"-"+title,netease_id=netease_id)
+                    curdate = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+                    media_sql = f'update media set netease_id="{netease_id}",artist_name="{ar_name}",image_path="{img_path}",lyric_path="{lrc_path}",update_at="{curdate}" where id="{mid}";'
+                    artist_sql = f'insert or ignore into artist(name) values("{ar_name}");'
+                    sql_con.sql2commit(media_sql)
+                    sql_con.sql2commit(artist_sql)
+                else:
+                    logger.warning(mid+"  "+title+" not match")
             except:
                 logger.error(format_exc())
         info_list_cp = info_list.copy()
-        sem = asyncio.Semaphore(5)
+        # sem = asyncio.Semaphore(5)
         while info_list:
-            loop = asyncio.get_event_loop()
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
             tasks = []
             for i in info_list_cp[:50]:
                 info_list.remove(i)
                 title = i.get("title")
                 album = i.get("album")
                 url = self.base_url + f"/cloudsearch?keywords={title} {album}&offset=0&limit=100"
-                tasks.append(asyncio.ensure_future(self.get_info(info=i,url=url,sem=sem)))
+                tasks.append(asyncio.ensure_future(self.get_info(info=i,url=url)))
             loop.run_until_complete(asyncio.wait(tasks))
             for r in tasks:
                 info,res = r.result()
@@ -99,7 +108,7 @@ class InfoCompletion:
     # 音乐信息刮削，目前支持网易云音乐
     def _completion(self):
         curdate = datetime.now().strftime("%Y-%m-%d")
-        logger.add(getcwd(),"log",f"info_completion_{curdate}.log")
+        logger.add(path.join(getcwd(),"log",f"info_completion_{curdate}.log"))
         logger.info("信息收集开始")
         try:
             yaml_conf = YamlConfig()
@@ -121,19 +130,19 @@ class InfoCompletion:
     def regular_start_completion(self):
         def start_fun():
             while True:
-                curtime = datetime.now("%H:%M")
+                curtime = datetime.now().strftime("%H:%M")
                 if curtime == "01:30":
                     self._completion()
                 time.sleep(60)
-        t = Thread(target=start_fun)
-        t.setDaemon(True)
-        t.start()
+        p = Thread(target=start_fun)
+        p.setDaemon(True)
+        p.start()
     # 手动启动刮削功能
     def start_completion(self):
         try:
-            t = Thread(target=self._completion)
-            t.setDaemon(True)
-            t.start()
+            p = Thread(target=self._completion)
+            p.setDaemon(True)
+            p.start()
             return INFO_COMPLETION_SUCCESS
         except:
             logger.error(format_exc())
